requirements.txt
flask
python-dotenv
pyjwt
pydantic
loguru
gunicorn
